Moved competition UI into own folder in preparation for WB UI
