+++ main_window/competition/mapping/map_data.py
