+++ main_window/competition/mapping/mapbox_utils.py
